back-end/requirements.txt
fastapi[all]
sqlalchemy
sqlmodel
email-validator
asyncpg
six
requests
numpy
websockets
PyJWT